backend/requirements.txt
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
gunicorn==21.2.0
requests==2.31.0